refactor(trace): split the binning helpers into bin_grid and batch_fold
